# Optimize Rack::Utils.select_best_encoding